# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText OpenAPI
  description: API for WASAText
  version: 1.0.0
security:
  - BearerAuth: [] 
...
paths:
...
  /conversations/{conversationId}/messages/{messageId}/comment:
    post:
      tags:
        - Reaction
      summary: Add a reaction to a specific message
      description: Adds a reaction or comment to a specified message within a conversation.
      operationId: commentMessage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        description: Payload containing the reaction or comment to add to the message.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reaction'
      responses:
        '204':
          description: Reaction added successfully     
        '400':
          description: Failed to add reaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Reaction
      summary: Remove a reaction from a specific message
      description: Removes a reaction or comment from a specified message within a conversation.
      operationId: uncommentMessage
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Reaction removed successfully
        '400':
          description: Failed to remove the reaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups:
    post:
      tags:
        - Group 
      summary: Create a group conversation
      description: |
        Creates a new conversation with type = "group" and adds the specified members (including the creator).
        The returned object is a Conversation; the `conversationId` acts as the group id.
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        description: Details required to create a group, including the group name and members.
        content:
          application/json:
            schema:
              type: object
              required: [groupName, members]
              properties:
                groupName:
                  type: string
                  description: Name of the group
                  minLength: 3
                  maxLength: 50
                members:
                  type: array
                  description: Array of user IDs to include
                  items:
                    type: string
                groupPhoto:
                  type: string
                  format: byte
                  description: Optional base64-encoded image data
      responses:
          '201':
            description: Group conversation created successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Conversation'
          '400':
            description: Failed to create group
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /groups/{groupId}:
    post:
      tags:
        - Group
      summary: Add a user to a group
      description: Allows an existing group member to add a user to the group. Here `groupId` equals the `conversationId`.
      operationId: addToGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      requestBody:
        required: true
        description: The username of the user to be added to the group.
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  $ref: '#/components/schemas/Username'
      responses:
        '204':
          description: User successfully added to the group
        '400':
          description: Failed to add user to the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Group
      summary: User leaves a group
      description: Allows a user to leave a specific group conversation. Here `groupId` equals the `conversationId`.
      operationId: leaveGroup
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            description: Unique identifier for the group conversation.
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      responses:
        '204':
          description: User successfully removed from group
        '400':
          description: Failed to remove the user from the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/name:
    put:
      tags:
        - Group 
      summary: Update a group name
      description: Updates the name of a specified group conversation (conversation type = group).
      operationId: setGroupName
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      requestBody:
        description: The new name for the group conversation.
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newName]
              properties:
                newName:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Group name successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: Failed to update group name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/photo: 
    put:
      tags:
        - Group 
      summary: Update group photo
      description: Updates the photo of a specified group conversation (conversation type = group).
      operationId: setGroupPhoto
      security:
        - BearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            pattern: ^.*?$
            minLength: 1
            maxLength: 36
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupPhoto]
              properties:
                groupPhoto:
                  type: string
                  format: byte
                  description: Base64-encoded image data
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Object that contains message
                required:
                  - message
                  - groupPhoto
                properties:
                  message:
                    type: string
                    description: Confirmation message
                  groupPhoto:
                    type: string
                    format: byte
                    description: Base64-encoded image data
                  
        '400':
          description: Failed to update group photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

       
#...
components:
  securitySchemes:
      BearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: |
          Use a valid JWT token in the Authorization header to access protected endpoints.
  schemas:
...
    Username:
      type: string
      pattern: ^[a-zA-Z0-9_]+$
      minLength: 3
      maxLength: 16
      description: A username containing only letters, numbers, or underscores, with a maximum length of 16 characters
...
    Conversation:
      type: object
      description: Represents a user's conversation.
      properties:
        conversationId:
          type: string
          description: Unique identifier for the conversation.
          pattern: ^.*?$
          minLength: 1
          maxLength: 36
        displayName:
          type: string
          description: Display name of the conversation (e.g., group name or other user's name).
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [direct, group]
          description: Indicates whether the conversation is a private chat or a group conversation.
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the conversation was created.
        members:
          type: array
          description: List of usernames in the conversation.
          items:
            $ref: '#/components/schemas/Username'
          minItems: 2
          maxItems: 100
        groupPhoto:
          type: string
          format: byte
          description: Base64-encoded image data
        lastMessage:
          $ref: '#/components/schemas/Message'
        messages:
          type: array
          description: List of messages in the conversation.
          items:
            $ref: '#/components/schemas/Message'
          minItems: 0
          maxItems: 100
    Message:
      type: object
      description: A message sent in a conversation.
      required:
        - id
        - sender
        - content
        - timestamp
        - message_status
        - reaction
        - attachment
      properties:
        id:
          type: string
          description: Unique identifier for the message.
        sender:
          type: object
          description: Information about the message sender.
          required:
            - id
            - username
          properties:
            id:
              type: string
              description: Unique identifier of the sender.
            username:
              $ref: '#/components/schemas/Username'
        content:
          type: object
          description: Content of the message.
          required:
            - type
            - value
          properties:
            type: 
              type: string
              enum: ['text', 'photo']
              description: Type of message (text or photo).
            value:
              type: string
              description: Message text or photo URL.
              pattern: ^.*?$
              maxLength: 500
              minLength: 1
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the message was sent.
        message_status:
          type: string
          enum: ['sent', 'delivered', 'read']
          description: Status of the message.
        reaction:
          type: array
          description: Reactions to this message.
          items:
            $ref: '#/components/schemas/Reaction'
          minItems: 0
          maxItems: 100
        attachments:
          type: array
          description: List of attachments.
          items:
            type: string
            format: byte
        forwarded_from:
          type: string
          description: ID of the user this message was forwarded from.
          pattern: ^.*?$
          minLength: 1
          maxLength: 36
    Reaction:
      type: object
      description: A reaction to a message.
      properties:
        emoji:
          type: string
          description: Emoji used for the reaction.
          pattern: ^.*?$
          minLength: 1
          maxLength: 10
        username:
          $ref: '#/components/schemas/Username'
...
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error details about why the request failed
...